reorganized and cleaned check syntax
