+++ include/lexer.h
@@ -6,7 +6,7 @@
 /*   By: nzhuzhle <nzhuzhle@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/15 19:58:38 by plinscho          #+#    #+#             */
-/*   Updated: 2023/12/21 19:17:10 by nzhuzhle         ###   ########.fr       */
+/*   Updated: 2023/12/22 15:53:33 by nzhuzhle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -54,9 +54,13 @@ int		len_no_q(char *s, char q, int len, int i);
 int		open_q(t_exp *exp, char c, int type);
 int		exp_quotes(t_mini *sh, t_lexer **head, int *flag);
 
-/***** check_syntax.c - *****/
+/***** check_syntax.c - checks the syntax *****/
 int	    check_input(char *in); // checks if there is anything in the input
 int		pre_quotes(char *line);
+int		check_syntax(t_mini *sh, t_lexer *current, int prev_token);
+char 	int_to_char(int num);
+int		ft_isspace(int c);
+/***************************************************/
 
 /***** utils.c - general utils *****/
 char	*ft_substr_quotes(char *s, char q, int len, int i); //check if it trims slashes like bash
@@ -64,16 +68,6 @@ int		check_chr(char c);
 char	**arr_clean(char **cmd, int flag); //frees a double array, if flag=0 - frees all the strings in it, if flag=1 only equals them to NULL (they are not allocated)
 int		ft_longer(char *str, char *key); // receives 2 strings and returns the lenth of the longer one
 char	*ft_smart_join(char *s1, char *s2, char *s3); // clean strjoin, that can jpoin 3 str
-/***************************************************/ 
-
-/***** check_sequence.c - checks the syntax sequence based on the tokens ***/
-//int		w_syntax(t_mini *sh);
-//void	check_sequence(t_mini *sh, char *seq);
-int		no_cmd(char *seq);
-//int		syntax_handler(t_lexer *head, int *pipes, int *redirs);
-int		check_syntax(t_mini *sh, t_lexer *current, int prev_token);
-char 	int_to_char(int num); //converts int into char
-int		ft_isspace(int c); // Detects if "c" is a space char.
 /***************************************************/
 
 //###########################################################################################
